packages/finnikacc-api/pyproject.toml
[project]
name = "finnikacc-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ihor", email = "Ihor.Halaniuk@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
]

[project.scripts]
finnikacc-api = "finnikacc_api:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = []
